feature(analytics): [PLTDA-1275] Round some of the Average metrics. (#1916)

+++ src/pages/stats/ClosedTicketsCellSummary.tsx
@@ -31,7 +31,7 @@ export const ClosedTicketsCellSummary = () => {
             ) : (
                 formatMetricValue(
                     metricValue,
-                    'decimal',
+                    'integer',
                     NOT_AVAILABLE_PLACEHOLDER
                 )
             )}
