+++ src/pages/stats/TicketsRepliedCellSummary.tsx
@@ -30,7 +30,7 @@ export const TicketsRepliedCellSummary = () => {
             ) : (
                 formatMetricValue(
                     metricValue,
-                    'decimal',
+                    'integer',
                     NOT_AVAILABLE_PLACEHOLDER
                 )
             )}
